chore: Store DeviceImpls as a Arc in ButtplugDevice
In order to generate our calls to devices to be lazy and only evaluate
when .await is used, we need to be able to add devices to our future.
All device functions take &self, so we can just pass along
Arc<DeviceImpl> and don't have to worry about lock conflicts.